% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postit_helpers.R
\docType{package}
\name{nidapFunctions-package}
\alias{nidapFunctions}
\alias{nidapFunctions-package}
\title{Internal and Exported Utilities for Post-it Rendering}
\description{
These functions support layout, coloring, sizing, and rendering of post-it
notes using grid and ggplot2 graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NIDAP-Community/nidapFunctions}
  \item Report bugs at \url{https://github.com/NIDAP-Community/nidapFunctions/issues}
}

}
\author{
\strong{Maintainer}: Cam Maggie \email{maggie.cam@nih.gov}

}
\keyword{internal}
