% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-packages.R
\name{nidapLoadPackages}
\alias{nidapLoadPackages}
\title{Load Packages in NIDAP}
\usage{
nidapLoadPackages(current.packages, suppress = TRUE)
}
\arguments{
\item{current.packages}{Vector of packages that are being used in the template}

\item{suppress}{Boolean to suppress messages.  By default set to TRUE}
}
\description{
This function will load packages and reorder package priorities
while suppressing startup messages by default.
}
\details{
This method provides a way to automatically (optionally) suppress
package start up messages in NIDAP and removes all other unnecessary
packages
}
