% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postit_helpers.R
\docType{data}
\name{postit_palette}
\alias{postit_palette}
\title{Named Color Palette}
\format{
Named character vector.
}
\usage{
postit_palette
}
\description{
A named character vector containing predefined hex color codes.
}
\keyword{internal}
