% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postit_helpers.R
\docType{data}
\name{postit_palette}
\alias{postit_palette}
\title{Color palette used for Post-it rendering}
\format{
Named character vector of hex colors.
}
\usage{
postit_palette
}
\description{
A named character vector of colors used in post-it visuals.
}
\keyword{internal}
