% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postit_helpers.R
\name{show_postit_palette}
\alias{show_postit_palette}
\title{Display Named Post-it Color Palette}
\usage{
show_postit_palette()
}
\value{
A ggplot2 object displaying the palette grid.
}
\description{
Visualizes all available named colors in the \code{postit_palette}, with labels and
hex codes.
}
\examples{
show_postit_palette()
}
